refactor: imageCollection validation
