+++ src/actions/images/collections/create.ts
@@ -2,13 +2,13 @@
 import { createZodActionError } from '@/actions/error'
 import { createImageCollection } from '@/server/images/collections/create'
 import { safeServerCall } from '@/actions/safeServerCall'
-import { createImageCollectionValidation } from '@/server/images/collections/schema'
+import { createImageCollectionValidation } from '@/server/images/collections/validation'
 import type { ImageCollection } from '@prisma/client'
 import type { ActionReturn } from '@/actions/Types'
-import type { CreateImageCollectionType } from '@/server/images/collections/schema'
+import type { CreateImageCollectionTypes } from '@/server/images/collections/validation'
 
 export async function createImageCollectionAction(
-    rawdata: FormData | CreateImageCollectionType
+    rawdata: FormData | CreateImageCollectionTypes['Type']
 ): Promise<ActionReturn<ImageCollection>> {
     const parse = createImageCollectionValidation.typeValidate(rawdata)
     if (!parse.success) return createZodActionError(parse)
