+++ src/server/images/collections/create.ts
@@ -1,12 +1,12 @@
 import 'server-only'
-import { createImageCollectionValidation } from './schema'
+import { createImageCollectionValidation } from './validation'
 import prisma from '@/prisma'
 import { prismaCall } from '@/server/prismaCall'
-import type { CreateImageCollectionType } from './schema'
+import type { CreateImageCollectionTypes } from './validation'
 import type { ImageCollection } from '@prisma/client'
 
 export async function createImageCollection(
-    rawdata: CreateImageCollectionType
+    rawdata: CreateImageCollectionTypes['Detailed']
 ): Promise<ImageCollection> {
     const data = createImageCollectionValidation.detailedValidate(rawdata)
     return await prismaCall(() => prisma.imageCollection.create({ data }))
